style: rename file for linting
