+++ packages/euclid/src/pool/pool_functions.rs
